Moving files and update account on login
